Fix test inheritance
See https://github.com/element-hq/synapse/pull/17167#discussion_r1594517041
